backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "pydantic[email]>=2.8.0",
    "minio>=7.2.10",
    "alembic>=1.16.4",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt[crypto]>=2.9.0",
    "email-validator>=2.2.0",
]
